Workaround for Anchor
TL;DR:  HTML code is parsed if it is inside a pre tag, regardless of
whether you use a HTML entity character instead.

So instead of

<pre>
&lt;foo&gt;bar&lt;/foo&gt;
</pre>

you get

<foo>bar</foo>

This is bad. The workaround is to use markdown to mark it as code, with
four leading spaces. Anchor, however, when wraps the text incorrectly
when it does this:

<p>
<code>
<foo>bar</foo>
</p>

It is now escaped, but not in a <pre> code block. I've increased the
font-size of <code> to compensate.